+++ examples/test_service/plugins/__init__.py
@@ -6,6 +6,8 @@
     :license: Inmanta EULA
 """
 
+# import a Python requirement to verify project installation
+import lorem  # noqa: F401
 from inmanta import resources
 from inmanta.agent import handler