+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 1.4.0 (?)
 Changes in this release:
+- Compatibility with `inmanta-service-orchestrator>=5`
 
 # v 1.3.0 (2021-09-23)
 Changes in this release: